+++ indexer/postgres/module.go
@@ -12,7 +12,7 @@ type ModuleIndexer struct {
 	moduleName   string
 	schema       schema.ModuleSchema
 	tables       map[string]*ObjectIndexer
-	definedEnums map[string]schema.EnumDefinition
+	definedEnums map[string]schema.EnumType
 	options      Options
 }
 
@@ -22,7 +22,7 @@ func NewModuleIndexer(moduleName string, modSchema schema.ModuleSchema, options 
 		moduleName:   moduleName,
 		schema:       modSchema,
 		tables:       map[string]*ObjectIndexer{},
-		definedEnums: map[string]schema.EnumDefinition{},
+		definedEnums: map[string]schema.EnumType{},
 		options:      options,
 	}
 }
@@ -30,29 +30,27 @@ func NewModuleIndexer(moduleName string, modSchema schema.ModuleSchema, options 
 // InitializeSchema creates tables for all object types in the module schema and creates enum types.
 func (m *ModuleIndexer) InitializeSchema(ctx context.Context, conn DBConn) error {
 	// create enum types
-	for _, typ := range m.schema.ObjectTypes {
-		err := m.createEnumTypesForFields(ctx, conn, typ.KeyFields)
-		if err != nil {
-			return err
-		}
-
-		err = m.createEnumTypesForFields(ctx, conn, typ.ValueFields)
-		if err != nil {
-			return err
-		}
+	var err error
+	m.schema.EnumTypes(func(enumType schema.EnumType) bool {
+		err = m.CreateEnumType(ctx, conn, enumType)
+		return err == nil
+	})
+	if err != nil {
+		return err
 	}
 
 	// create tables for all object types
-	for _, typ := range m.schema.ObjectTypes {
+	m.schema.ObjectTypes(func(typ schema.ObjectType) bool {
 		tm := NewObjectIndexer(m.moduleName, typ, m.options)
 		m.tables[typ.Name] = tm
-		err := tm.CreateTable(ctx, conn)
+		err = tm.CreateTable(ctx, conn)
 		if err != nil {
-			return fmt.Errorf("failed to create table for %s in module %s: %v", typ.Name, m.moduleName, err) //nolint:errorlint // using %v for go 1.12 compat
+			err = fmt.Errorf("failed to create table for %s in module %s: %v", typ.Name, m.moduleName, err) //nolint:errorlint // using %v for go 1.12 compat
 		}
-	}
+		return err == nil
+	})
 
-	return nil
+	return err
 }
 
 // ObjectIndexers returns the object indexers for the module.