+++ testutil/network/network.go
@@ -163,11 +163,7 @@ func DefaultConfig(factory TestFixtureFactory) Config {
 	}
 }
 
-func DefaultConfigWithAppConfig(appConfig depinject.Config) (Config, error) {
-	return DefaultConfigWithAppConfigWithQueryGasLimit(appConfig, 0)
-}
-
-func DefaultConfigWithAppConfigWithQueryGasLimit(appConfig depinject.Config, queryGasLimit uint64) (Config, error) {
+func DefaultConfigWithAppConfig(appConfig depinject.Config, baseappOpts ...func(*baseapp.BaseApp)) (Config, error) {
 	var (
 		appBuilder            *runtime.AppBuilder
 		txConfig              client.TxConfig
@@ -222,10 +218,11 @@ func DefaultConfigWithAppConfigWithQueryGasLimit(appConfig depinject.Config, que
 		app := appBuilder.Build(
 			dbm.NewMemDB(),
 			nil,
-			baseapp.SetPruning(pruningtypes.NewPruningOptionsFromString(val.GetAppConfig().Pruning)),
-			baseapp.SetMinGasPrices(val.GetAppConfig().MinGasPrices),
-			baseapp.SetChainID(cfg.ChainID),
-			baseapp.SetQueryGasLimit(queryGasLimit),
+			append(baseappOpts,
+				baseapp.SetPruning(pruningtypes.NewPruningOptionsFromString(val.GetAppConfig().Pruning)),
+				baseapp.SetMinGasPrices(val.GetAppConfig().MinGasPrices),
+				baseapp.SetChainID(cfg.ChainID),
+			)...,
 		)
 
 		testdata.RegisterQueryServer(app.GRPCQueryRouter(), testdata.QueryImpl{})
@@ -350,7 +347,7 @@ func New(l Logger, baseDir string, cfg Config) (NetworkI, error) {
 				apiListenAddr = cfg.APIAddress
 			} else {
 				if len(portPool) == 0 {
-					return nil, fmt.Errorf("failed to get port for API server")
+					return nil, errors.New("failed to get port for API server")
 				}
 				port := <-portPool
 				apiListenAddr = fmt.Sprintf("tcp://127.0.0.1:%s", port)
@@ -367,7 +364,7 @@ func New(l Logger, baseDir string, cfg Config) (NetworkI, error) {
 				cmtCfg.RPC.ListenAddress = cfg.RPCAddress
 			} else {
 				if len(portPool) == 0 {
-					return nil, fmt.Errorf("failed to get port for RPC server")
+					return nil, errors.New("failed to get port for RPC server")
 				}
 				port := <-portPool
 				cmtCfg.RPC.ListenAddress = fmt.Sprintf("tcp://127.0.0.1:%s", port)
@@ -377,7 +374,7 @@ func New(l Logger, baseDir string, cfg Config) (NetworkI, error) {
 				appCfg.GRPC.Address = cfg.GRPCAddress
 			} else {
 				if len(portPool) == 0 {
-					return nil, fmt.Errorf("failed to get port for GRPC server")
+					return nil, errors.New("failed to get port for GRPC server")
 				}
 				port := <-portPool
 				appCfg.GRPC.Address = fmt.Sprintf("127.0.0.1:%s", port)
@@ -410,14 +407,14 @@ func New(l Logger, baseDir string, cfg Config) (NetworkI, error) {
 		monikers[i] = nodeDirName
 
 		if len(portPool) == 0 {
-			return nil, fmt.Errorf("failed to get port for Proxy server")
+			return nil, errors.New("failed to get port for Proxy server")
 		}
 		port := <-portPool
 		proxyAddr := fmt.Sprintf("tcp://127.0.0.1:%s", port)
 		cmtCfg.ProxyApp = proxyAddr
 
 		if len(portPool) == 0 {
-			return nil, fmt.Errorf("failed to get port for Proxy server")
+			return nil, errors.New("failed to get port for Proxy server")
 		}
 		port = <-portPool
 		p2pAddr := fmt.Sprintf("tcp://127.0.0.1:%s", port)